backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.135.2",
    "mss>=10.1.0",
    "openai>=2.30.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.42.0",
]
